ref(hybrid-cloud): Combines internal integration exception files
