+++ src/sentry/integrations/gitlab/blame.py
@@ -15,8 +15,7 @@ from sentry.integrations.gitlab.utils import (
 )
 from sentry.integrations.mixins.commit_context import CommitInfo, FileBlameInfo, SourceLineInfo
 from sentry.shared_integrations.client.base import BaseApiClient
-from sentry.shared_integrations.exceptions import ApiRateLimitedError
-from sentry.shared_integrations.exceptions.base import ApiError
+from sentry.shared_integrations.exceptions import ApiError, ApiRateLimitedError
 from sentry.shared_integrations.response.sequence import SequenceApiResponse
 from sentry.utils import json, metrics
