+++ src/sentry/shared_integrations/exceptions/__init__.py
@@ -4,21 +4,88 @@ import errno
 from typing import Any, Mapping
 from urllib.parse import urlparse
 
+from bs4 import BeautifulSoup
+from requests import Response
 from requests.exceptions import RequestException
-from rest_framework.request import Request
-from rest_framework.response import Response
+from rest_framework.request import Request as RestFrameworkRequest
+from rest_framework.response import Response as RestFrameworkResponse
 
-from .base import ApiError
+from sentry.utils import json
 
 __all__ = (
     "ApiConnectionResetError",
     "ApiError",
+    "ApiConflictError",
     "ApiHostError",
     "ApiTimeoutError",
     "ApiUnauthorized",
     "ApiRateLimitedError",
     "IntegrationError",
 )
+
+
+class ApiError(Exception):
+    """
+    Base class for errors which arise while making outgoing requests to third-party APIs.
+    """
+
+    code: int | None = None
+
+    def __init__(
+        self,
+        text: str,
+        code: int | None = None,
+        url: str | None = None,
+        host: str | None = None,
+        path: str | None = None,
+    ) -> None:
+        if code is not None:
+            self.code = code
+        self.text = text
+        self.url = url
+        # we allow `host` and `path` to be passed in separately from `url` in case
+        # either one is all we have
+        self.host = host
+        self.path = path
+        self.json: dict[str, Any] | None = None
+        self.xml: BeautifulSoup | None = None
+
+        # TODO(dcramer): pull in XML support from Jira
+        if text:
+            try:
+                self.json = json.loads(text)
+            except (json.JSONDecodeError, ValueError):
+                if self.text[:5] == "<?xml":
+                    # perhaps it's XML?
+                    self.xml = BeautifulSoup(self.text, "xml")
+
+        if url and not self.host:
+            try:
+                self.host = urlparse(url).netloc
+            except ValueError:
+                self.host = "[invalid URL]"
+
+        if url and not self.path:
+            try:
+                self.path = urlparse(url).path
+            except ValueError:
+                self.path = "[invalid URL]"
+
+        super().__init__(text[:1024])
+
+    def __str__(self) -> str:
+        return self.text
+
+    @classmethod
+    def from_response(cls, response: Response, url: str | None = None) -> ApiError:
+        if response.status_code == 401:
+            return ApiUnauthorized(response.text, url=url)
+        elif response.status_code == 429:
+            return ApiRateLimitedError(response.text, url=url)
+        elif response.status_code == 409:
+            return ApiConflictError(response.text, url=url)
+
+        return cls(response.text, response.status_code, url=url)
 
 
 class ApiHostError(ApiError):
@@ -31,7 +98,7 @@ class ApiHostError(ApiError):
         return cls("Unable to reach host")
 
     @classmethod
-    def from_request(cls, request: Request) -> ApiHostError:
+    def from_request(cls, request: RestFrameworkRequest) -> ApiHostError:
         host = urlparse(request.url).netloc
         return cls(f"Unable to reach host: {host}", url=request.url)
 
@@ -46,7 +113,7 @@ class ApiTimeoutError(ApiError):
         return cls("Timed out reaching host")
 
     @classmethod
-    def from_request(cls, request: Request) -> ApiTimeoutError:
+    def from_request(cls, request: RestFrameworkRequest) -> ApiTimeoutError:
         host = urlparse(request.url).netloc
         return cls(f"Timed out attempting to reach host: {host}", url=request.url)
 
@@ -96,6 +163,8 @@ class IntegrationFormError(IntegrationError):
 class ClientError(RequestException):
     """4xx Error Occurred"""
 
-    def __init__(self, status_code: str, url: str, response: Response | None = None) -> None:
+    def __init__(
+        self, status_code: str, url: str, response: RestFrameworkResponse | None = None
+    ) -> None:
         http_error_msg = f"{status_code} Client Error: for url: {url}"
         super().__init__(http_error_msg, response=response)
