+++ tests/sentry/integrations/jira/test_webhooks.py
@@ -15,7 +15,7 @@ from sentry.integrations.utils import AtlassianConnectValidationError
 from sentry.models.integrations.integration import Integration
 from sentry.services.hybrid_cloud.integration.serial import serialize_integration
 from sentry.services.hybrid_cloud.organization.serial import serialize_rpc_organization
-from sentry.shared_integrations.exceptions.base import ApiError
+from sentry.shared_integrations.exceptions import ApiError
 from sentry.testutils.cases import APITestCase, TestCase
 
 TOKEN = "JWT anexampletoken"
