+++ tests/sentry/tasks/integrations/github/test_open_pr_comment.py
@@ -7,7 +7,7 @@ from django.utils import timezone
 from sentry.models.group import Group, GroupStatus
 from sentry.models.options.organization_option import OrganizationOption
 from sentry.models.pullrequest import CommentType, PullRequest, PullRequestComment
-from sentry.shared_integrations.exceptions.base import ApiError
+from sentry.shared_integrations.exceptions import ApiError
 from sentry.tasks.integrations.github.open_pr_comment import (
     format_issue_table,
     format_open_pr_comment,
